+++ site/src/other.rs
@@ -570,7 +570,6 @@ pub fn Subscripts() -> impl IntoView {
         subscript(Repeat, "Repetition count", "⍥₅(⊂⟜/+) [1 2]"),
         subscript(Tuples, "Tuple size", "⧅₂< ⇡4"),
         subscript(Stencil, "Window size", "⧈₃∘ ⇡6"),
-        subscript(Map, "Preallocation size", "˙map₃ [3 2 1]"),
         subscript(Args, "Print top N values", "?₂ 1 2 3 4"),
     ];